⚡️ :: Integer -> Double로 수정

+++ src/main/java/com/walkhub/walkhub/domain/user/domain/CalorieLevel.java
@@ -1,54 +1,54 @@
 package com.walkhub.walkhub.domain.user.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
+
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 public class CalorieLevel {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-	@NotNull
-	@Length(max = 10)
-	private String foodName;
+    @NotNull
+    @Length(max = 10)
+    private String foodName;
 
-	@NotNull
-	private Integer size;
+    @NotNull
+    private Integer size;
 
-	@NotNull
-	private String message;
+    @NotNull
+    private String message;
 
-	@NotNull
-	private String foodImageUrl;
+    @NotNull
+    private String foodImageUrl;
 
-	@NotNull
-	@Column(columnDefinition = "TINYINT")
-	private Integer level;
+    @NotNull
+    @Column(columnDefinition = "TINYINT")
+    private Integer level;
 
-	@NotNull
-	private Integer calorie;
+    @NotNull
+    private Double calorie;
 
-	@Builder
-	public CalorieLevel(String foodName, Integer size, String message, String foodImageUrl, Integer level,
-		Integer calorie) {
-		this.foodName = foodName;
-		this.size = size;
-		this.message = message;
-		this.foodImageUrl = foodImageUrl;
-		this.level = level;
-		this.calorie = calorie;
-	}
+    @Builder
+    public CalorieLevel(String foodName, Integer size, String message, String foodImageUrl, Integer level, Double calorie) {
+        this.foodName = foodName;
+        this.size = size;
+        this.message = message;
+        this.foodImageUrl = foodImageUrl;
+        this.level = level;
+        this.calorie = calorie;
+    }
 }
